# fix: sanatize structured metadata at query time