@rfratto 2024-08-27
Will this break anything for keys coming from OTel attributes?