@MasslessParticle 2024-08-27
@rfratto yes and no. Yes because Loki will happily ingest periods so users will be querying for something unexpected. No because this data is unqueryable right now the invalid labels will break [here](https://github.com/grafana/loki/blob/1eb8342d41a9fdb6c5fcd3e6e5a8c6b98bde4e43/pkg/iter/categorized_labels_iterator.go#L39).